requirements.txt
aiofiles==23.2.1
beautifulsoup4==4.12.3
boto3==1.34.128
botocore==1.34.128
colorama==0.4.6
greenlet==3.0.3
jmespath==1.0.1
numpy==1.26.4
pandas==2.2.2
pathlib==1.0.1
playwright==1.44.0
pyee==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
